+++ Source/Gapotchenko.GnuTK/Diagnostics/DiagnosticServices.cs
@@ -1,0 +1,9 @@
+﻿namespace Gapotchenko.GnuTK.Diagnostics;
+
+static class DiagnosticServices
+{
+    public static string GetErrorIdentifier(DiagnosticCode? errorCode)
+    {
+        return $"GNUTK{(int?)errorCode:D4}";
+    }
+}